+++ src/GRA.Controllers/ViewModel/MissionControl/Participants/ParticipantsDetailViewModel.cs
@@ -6,11 +6,13 @@ namespace GRA.Controllers.ViewModel.MissionControl.Participants
     {
         public GRA.Domain.Model.User User { get; set; }
         public int Id { get; set; }
+        public string Username { get; set; }
         public int HouseholdCount { get; set; }
         public int PrizeCount { get; set; }
         public int? HeadOfHouseholdId { get; set; }
         public bool HasAccount { get; set; }
         public bool CanEditDetails { get; set; }
+        public bool CanEditUsername { get; set; }
         public bool RequirePostalCode { get; set; }
         public bool ShowAge { get; set; }
         public bool ShowSchool { get; set; }